# docs/api/openapi.json
{
  "info": {
    "description": "Generated route inventory for the local-only gateway. Typed request and response schemas remain documented in rest-api.md until they are migrated into this contract.",
    "title": "AIPermission Local REST API",
    "version": "route-inventory-v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/backup/providers/{id}": {
      "delete": {
        "operationId": "deleteApiBackupProvidersById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "See the REST API reference for status and response body details."
          }
        },
        "tags": [
          "backup"
        ],
        "x-aipermission-contract-level": "route-inventory"
      },
      "put": {
        "operationId": "putApiBackupProvidersById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "See the REST API reference for status and response body details."
          }
        },
        "tags": [
          "backup"
        ],
        "x-aipermission-contract-level": "route-inventory"
      }
    },
...
    "/api/connector-targets/{id}/profiles/{profile_id}/backup": {
      "get": {
        "operationId": "getApiConnectorTargetsByIdProfilesByProfileIdBackup",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "See the REST API reference for status and response body details."
          }
        },
        "tags": [
          "connector-targets"
        ],
        "x-aipermission-contract-level": "route-inventory"
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3210"
    }
  ],
...
}